requirements.txt
python-telegram-bot==20.3
telethon==1.28.5
SQLAlchemy==2.0.15
python-dotenv==1.0.0
pytz==2023.3